pyproject.toml
[tool.poetry]
name = "cpp-msgpack-rpc"
version = "0.0.0"
description = ""
authors = ["Kenta Kabashima <kenta_program37@hotmail.co.jp>"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
# Some tools implicitly require setuptools.
setuptools = "^72.1.0"

[tool.poetry.group.doc.dependencies]
myst-parser = "^4.0.0"
sphinx = "^7.3.7"
sphinx-autobuild = "^2024.4.16"
sphinxcontrib-plantuml = "^0.30"
sphinx-orange-book-theme = "^0.11.0"
breathe = "^4.35.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-approvaltests = "^0.2.4"
msgpack = "^1.0.8"
pytest-asyncio = "^0.23.8"
psutil = "^6.0.0"
pandas = "^2.2.2"
plotly = "^5.23.0"
kaleido = "==0.2.1"
statsmodels = "^0.14.2"
fastjsonschema = "^2.20.0"
pytomlpp = "^1.0.13"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
click = "^8.1.7"
cmakelang = "^0.6.13"
doc8 = "^1.1.1"
pre-commit = "^3.8.0"
clang-tidy-checker = "^0.5.0"
lcov_cobertura = "^2.0.2"
tqdm = "^4.66.5"
lizard = "^1.17.10"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.main]
disable = "too-many-instance-attributes,line-too-long"
